requirements.txt
streamlit==1.32.0
streamlit-option-menu==0.3.6
pandas==2.1.3
numpy==1.26.2
pillow==10.2.0
